fix: refactor settings location for notifications
